vigil_upgrade/requirements.txt
# Vigil Upgrade — 2026-era pipeline (use from project root: pip install -r vigil_upgrade/requirements.txt)
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
PyYAML>=6.0
pandas>=2.0.0
scikit-learn>=1.3.0
# Optional: ReID (proactive package)
# torchreid  # OSNet; or use ResNet fallback in proactive.reid
